# src/openapi.yaml
openapi: 3.0.3
info:
  title: Blackbox API Proxy
  version: 1.0.0
  description: >
    A blazing-fast, lightweight reverse proxy written in Rust that provides a standard 
    OpenAI-compatible `/chat/completions` API interface to the internal Blackbox AI backend.
servers:
  - url: http://localhost:8080
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Your PROXY_API_KEY (defaults to `xyz`) or your Blackbox API token.
security:
  - bearerAuth: []
paths:
  /chat/completions:
    post:
      summary: Create chat completion
      description: Send a chat completion request to the Blackbox proxy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - model
                - messages
              properties:
                model:
                  type: string
                  example: minimax-m2.7
                  description: The model to route to (e.g. `minimax-m2.7`, `kimi-k2.6`).
                messages:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
                stream:
                  type: boolean
                  default: false
                tools:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        default: function
                      function:
                        type: object
      responses:
        '200':
          description: Chat completion response
        '401':
          description: Unauthorized - Invalid PROXY_API_KEY